Extract Load, Store, Call, and MethodBody emission into separate partial classes
